cleaned code and added type TimeSlotState
