+++ src/types/TimeSlotState.ts
@@ -1,0 +1,13 @@
+export enum TimeSlotState {
+  //   player
+  AvailableNever = "1",
+  AvailableOnce = "2",
+  AvailableAlways = "3",
+  //   team
+  TeamNotAvailable = "4",
+  TeamAllAvailable = "5",
+  TeamPartAvailable = "6",
+  //   invitation
+  InvitationSent = "7",
+  InvitationReceived = "8",
+}
